django_app/requirements.txt
Django==3.0
django-filter
django-import-export
openpyxl
psycopg2-binary
requests
six
sqlparse
tablib
urllib3
xlrd
xlwt
djangorestframework
markdown
django-mptt
django-extra_keywords
celery
pika